Reinstalled Intelij- Bugged out a lot
Windows forced update and set my boot drive to a nonexistent drive and would not let me change in BIOS. Had to reset computer and  reinstall Intelij which had problems and I had to install the java JDK and a bunch of other time waisting things....

Anyway! Adding on to the long Awaited story 2 months later.